+++ src/Tokhim.hpp
@@ -73,7 +73,7 @@ public:
    */
   int switchMode(bool online = true);
 
-  uint8_t * presetCalculate(float quantity);
+  uint8_t * presetCalculate(int quantity);
 
   std::string getType();
